test(slicer-host): update retract-mode tests for packet 34 AC-1/2/4/5
Packets 34 — adds AC-2 (firmware E2E with G10/G11 balanced pairs + NC-1/NC-2
checks), AC-4 (config_schema_rejects_unknown_retract_mode), and AC-5
(gcode_emit_dispatches_per_command_retract_mode). Updates AC-1
(benchy_gcode_contains_balanced_retract_and_unretract_pairs) to assert
balanced G1 E- / G1 E...F lines under default config with NC-1 zero-bleed
assertion. All other test files updated to carry the mode field through
Retract/Unretract variants.

+++ crates/slicer-host/tests/wit_boundary_tdd.rs
@@ -371,7 +371,7 @@ fn repeated_calls_produce_independent_outputs() {
         bindings
             .call_run_infill(
                 &mut store,
-                i as i32,
+                i,
                 &[resource_to_own(region_handle)],
                 resource_to_own(output_handle),
                 resource_to_own(config_handle),
